validate

+++ Electio/src/java/Action/ElectionResult.java
@@ -37,7 +37,7 @@ public class ElectionResult implements Controller.Action {
                 long election_id = Long.parseLong(elec_id);
                 DBDAOImplElection objE = DBDAOImplElection.getInstance();
                 DBDAOImplCandidate objC = DBDAOImplCandidate.getInstance();
-                int election_type = (int) objE.getElectionType(election_id).getType_id();
+                long election_type = (int) objE.getElectionType(election_id).getType_id();
                 req.setAttribute("election_type", election_type + "");
                 if (election_type == 1) {
                     candidates = objC.getCandidatesForPreferentialVoting(election_id);
